requirements.txt
# Core ML Libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# MLflow for experiment tracking
mlflow>=2.8.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Data validation
pydantic>=2.0.0

# Monitoring and Logging
prometheus-client>=0.15.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0

# Development tools
jupyter>=1.0.0
python-multipart>=0.0.5

# Plotting
matplotlib>=3.6.0
seaborn>=0.12.0

# Azure Dependencies
asyncpg>=0.28.0
azure-storage-blob>=12.17.0
azure-identity>=1.14.0
opencensus-ext-azure>=1.1.0
opencensus-ext-flask>=0.7.0
